requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
flask-mail
itsdangerous
python-dotenv
flasgger
requests
psycopg2-binary
gunicorn
flask-dance[google]
blinker
flask-wtf
flask-limiter
flask-bcrypt
